\name{StochaTR-package}
\alias{StochaTR-package}
\alias{StochaTR}
\docType{package}
\title{
Solving stochastic linear programs with a single risk constraint
}
\description{
Solving stochastic linear programs with a single risk constraint. Coherent distortion risk measures are used. 
}
\details{
\tabular{ll}{
Package: \tab StochaTR\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2011-10-15\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
Solving stochastic linear programs with a single risk constraint. Coherent distortion risk measures are used.  
}
\author{
Pavel Bazovkin, Karl Mosler

Maintainer: Pavel Bazovkin <econometrics@russia.ru> 
}
\references{
Pavel Bazovkin and Karl Mosler. An exact algorithm for weighted-mean trimmed regions in any dimension. 201x. Submitted.

Bazovkin, P. and Mosler, K. Stochastic linear programming with distortion risk constraints. 201x. Working paper.

}

\keyword{ package }
\keyword{ programming }
\keyword{ multivariate }
\keyword{ nonparametric }
\keyword{ models }

\examples{

ls()

}
